rework db access layer
