+++ blazor/Shell/Program.cs
@@ -1,5 +1,5 @@
 using Microsoft.EntityFrameworkCore;
-using Shell;
+using Shell.Db;
 using Shell.Components;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -7,7 +7,7 @@ var builder = WebApplication.CreateBuilder(args);
 var conString = builder.Configuration.GetConnectionString("PostgresConnection") ??
      throw new InvalidOperationException("Connection string 'PostgresConnection' not found.");
 
-builder.Services.AddDbContextFactory<BlazorDbContext>(options =>
+builder.Services.AddDbContextFactory<ShellDbContext>(options =>
     options.UseNpgsql(conString));
 
 // Add services to the container.
